checks.py, tests/unit/: move check.py to tests/ for usage in integration tests
